feat!: starting node(s) are now determined automatically. Each starting node gets its own rank, regardless of double entries
