+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@crinkles/digl",
-  "version": "1.0.0",
+  "version": "2.0.0",
   "description": "Autolayout graphs",
   "main": "dist/index.js",
   "module": "dist/esm/index.js",
